Refactoring, code breakout into smaller units.
